requirements.txt
py-clob-client>=0.34.0
websockets>=12.0
aiohttp>=3.9.0
rich>=13.7.0
python-dotenv>=1.0.0
